requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
et-xmlfile==1.1.0
fastapi==0.73.0
fuzzywuzzy==0.18.0
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
numpy==1.22.2
openpyxl==3.0.9
pandas==1.4.0
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-Levenshtein==0.12.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing-extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.4
uvloop==0.16.0
watchgod==0.7
websockets==10.1
